+++ src/lightning_lite/plugins/__init__.py
@@ -35,4 +35,5 @@ __all__ = [
     "Precision",
     "TPUPrecision",
     "TPUBf16Precision",
+    "FSDPPrecision",
 ]