+++ src/lightning_lite/plugins/precision/__init__.py
@@ -26,4 +26,5 @@ __all__ = [
     "Precision",
     "TPUPrecision",
     "TPUBf16Precision",
+    "FSDPPrecision",
 ]