+++ tests/tests_lite/strategies/test_fsdp.py
@@ -1,0 +1,53 @@
+# Copyright The PyTorch Lightning team.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+from unittest import mock
+
+import pytest
+from tests_lite.helpers.runif import RunIf
+
+from lightning_lite.strategies import FSDPStrategy
+from lightning_lite.utilities.imports import _TORCH_GREATER_EQUAL_1_12
+
+if _TORCH_GREATER_EQUAL_1_12:
+    from torch.distributed.fsdp.fully_sharded_data_parallel import MixedPrecision
+
+
+@mock.patch("lightning_lite.strategies.fsdp._TORCH_GREATER_EQUAL_1_12", False)
+def test_fsdp_support(*_):
+    with pytest.raises(NotImplementedError, match="`FSDPStrategy` is supported from PyTorch v1.12.0"):
+        FSDPStrategy()
+
+
+@RunIf(min_torch="1.12")
+# @mock.patch("lightning_lite.strategies.fsdp.FullyShardedDataParallel")
+def test_fsdp_custom_mixed_precision(*_):
+    """Test that passing a custom mixed precision config works."""
+    config = MixedPrecision()
+    strategy = FSDPStrategy(mixed_precision=config)
+    assert strategy.mixed_precision_config == config
+    #
+    #
+    # wrapped_module = strategy.setup_module(nn.Linear(3, 3))
+    # assert wrapped_module.mixed_precision == config
+
+
+def custom_auto_wrap_policy(
+    module,
+    recurse,
+    unwrapped_params: int,
+    min_num_params: int = int(1e8),
+) -> bool:
+    return unwrapped_params >= 2